New attribute "PAG statement head" is now added to control
New attribute "PAG statement head" is now added to control
statements (if, switch, loops). This points to the ICFG block
implementing the head of this statement, e.g. for a SgIfStmt it is
the block actually containing the if statement. This is used to
access outgoing true/false edges and their analysis information.

darcs-hash:20070412024200-3fb0b-78628174de628b7387bf647d23f8eb84803ff825.gz