Added support for a function expression and a parameter list in the
Added support for a function expression and a parameter list in the
ExternalCall node.

Improved the way the ICFG for logical expressions is built (we now make sure
that the logical variable is always assigned a boolean value), but the
evaluation order for right-associative expressions like a || (b && c) is
still broken.

darcs-hash:20071023152448-d8bd4-db7d51cdafe5ffbf05d64c3f72981c6707a99abd.gz