ELF symbols are added to the ExecHeader object.
If an ELF ".symtab" section is present then its symbols are
added to the ELF header object.  Otherwise if a ".dynsym"
section is present those symbols are added. Normally,
".dynsym" contains a subset of the information contained
by ".symtab". In any case, the full set of symbols from
both sections (when present) are stored along with their
section.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@768 a27c51de-2d2b-4d1b-a6e8-f7f87814f453