PE .idata ctor moved from segment->section; section RVAs
The constructor for the PE ".idata" object was moved from being a
subclass of ExecSegment to ExecSection in order to align better with
Elf (PE doesn't have sections or segments).

The sections generated for PE are assigned relative virtual addresses
(RVAs) that are the same as the segments since PE objects are mapped
to a generic sections aligned 1:1 with generic segments.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@816 a27c51de-2d2b-4d1b-a6e8-f7f87814f453