DOS Real-mode segment size limited for PE/LE/NE/LX formats
The DOS real-mode (text/data/etc.) section size is declared
in the DOS file header and usually extends well into the
PE/LE/NE/LX part of the executable for no apparent reason.
In order to better detect holes in the latter formats, the
size of the DOS real-mode section is limited so as not to
extend beyond the beginning of the PE/LE/NE/LX file header.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@1173 a27c51de-2d2b-4d1b-a6e8-f7f87814f453