ELF segment table merged with section table
Executable files have a segment table in addition to the
optional section table. The section table has more info and
usually a finer granularity, so we use that to define the
ExecSections of the file. However, when the segment table
is present we merge its contents with the section table.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@1333 a27c51de-2d2b-4d1b-a6e8-f7f87814f453