Calls to SgAsmGenericSection::content fixed.
First, these two are not equivalent:
  unsigned n = le_to_host((const uint16_t) relocs->content(at, 2)[0]);
  unsigned n = le_to_host(*(const uint16_t*)content(offset,2))
(The second version is the correct one. The first version discards
one of the bytes from the file before optionally swapping them.)

Second, removed redundant expressions now that content() is again
returning a ptr to mmap'd memory.

Third, places where result of content() was cast are working again
since we're not casting the SgUnsignedCharList object.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@1582 a27c51de-2d2b-4d1b-a6e8-f7f87814f453