Checkpoint (not compiling)
Reworked section constructors in anticipation of moving section lists
into SgAsmGenericHeader.  A side effect of this is that we generally
don't need to call SgAsmGenericSection::set_header() on any of the new
sections since this is now done in the SgAsmGenericSection constructor.

Also removed some already-commented-out set_parent() calls and removed
a couple of places where the function was called twice.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@1861 a27c51de-2d2b-4d1b-a6e8-f7f87814f453