ElfDynamicSection reworked.
This reduces the amount of code while increasing the
number of tag/value combinations that are recognized.

It also makes it possible to relocate sections to
which the dynamic section points and when the
dynamic section is unparsed the values will be based
on the new locations of the referenced sections.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@2334 a27c51de-2d2b-4d1b-a6e8-f7f87814f453