New type rose_rva_t similar in use to rose_addr_t
The main difference is that rose_rva_t is always
a virtual address and relative to either the file base
virtual address or the beginning of some section.

The main benefit is that by being associated with
a section, any time the section mapped address
changes (such as when some other section is resized
or shifted), all rose_rva_t objects associated with
that section are automatically adjusted as well.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@2338 a27c51de-2d2b-4d1b-a6e8-f7f87814f453