ELF Header p_e_phentsize -> p_phextrasz
A segment table entry has a required part and an optional part. The ELF header indicates the sum of
these two values. Internally, we now store only the size of the extra part since the size of the
required part can be calculated from the current word size.

This modification allows us to more easily change the word size.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@2363 a27c51de-2d2b-4d1b-a6e8-f7f87814f453