ExecFormats section reallocation modifications
The SgAsmGenericStrtab::reallocate() was generalized to
SgAsmGenericSection::reallocate() so that other things
besides string tables get an opportunity to resize
themselves before unparsing.

This fix makes it possible for the file format word
size to be changed in a header (such as the ELF File
Header) and that change get propogated to any section
that needs to make adjustments to its size.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@2376 a27c51de-2d2b-4d1b-a6e8-f7f87814f453