Fixed bug in ExecFormat tests
In src/midend/astDiagnostics/AstConsistencyTests.C
TestChildPointersInMemoryPool::visit(SgNode*node) we test that all
nodes appear as children of their parent. However, the logic also checks
nodes that are unreachable from the root of the AST (nodes that are
orphans) and thus if the parent pointer is dangling we get some sort of fault.

The fix was to make sure that whenever a node has an explicit destructor
(as opposed to the ROSETTA-generated destructor) that we also destroy
all child nodes.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@2404 a27c51de-2d2b-4d1b-a6e8-f7f87814f453