ELF Symbol unparsing careful about string table
When a symbol table is parsed, symbols that don't have names
but appear to represent an ELF Section are given the name of
the ELF Section. When unparsing we're careful not to store
the name, but keep it as it was originally.

User should beware that changing the name of such a symbol
with code like:  symbol->get_name()->set_string("foo")
will change the name of the section as well as the symbol.
If you want to assign the symbol a name without changing
the section name then use symbol->set_name(), supplying
a newly constructed name in the appropriate string table.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@2520 a27c51de-2d2b-4d1b-a6e8-f7f87814f453