PE Import Table (.idata) parsing is more robust.
In particular, it can handle cases when the Import Directory
Entry points to RVAs that are outside the Import Table as long
as those RVAs are mapped by some other section. This sometimes
happens with *.dll files where the RVA/Size pair for the Import
Table is too small and there is no ".idata" section in the PE
Section Table.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@2809 a27c51de-2d2b-4d1b-a6e8-f7f87814f453