PE Import Table is parsed
This rewrite of the parser is to make it more general and able
to handle cases where parts of the import information is outside
the address space assigned to the .idata section and/or the
Import Table.

Does not yet support unparsing (other than by writing out the
original bytes).

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@2945 a27c51de-2d2b-4d1b-a6e8-f7f87814f453