Unparser tries not to write past orig EOF
The various "tiny" executables use the trick that sections
can be defined to extend beyond EOF and the loader will
fill that part of the section with zeros. In order to make the
unparser generate a file identical to the original, the
unparser doesn't write trailing zeros that are past the
original EOF.

When modifying the AST to extend the file, if you want the
unparsed file to be extended to its full size then you
should call SgAsmGenericFile::set_orig_size() to a
suitably large value. (Non-zero bytes are *always* written.)

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@3116 a27c51de-2d2b-4d1b-a6e8-f7f87814f453