Allow ROSETTA to generate some functions
I keep going back and forth on this. On the one hand we want ROSETTA to
do this because they're such simple functions. On the other hand we
don't want users to call these particular ones after the AST is
generated because the unparser calculates these redundant things from
other data in the IR nodes.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@3331 a27c51de-2d2b-4d1b-a6e8-f7f87814f453