Update make check answer
This is due to a bug fix. The old code failed to set the file/memory
alignment constraints for ELF Segments based on values in the ELF
Segment Table. Since the new code sets these values, a call to
SgAsmGenericFile::shift_extend(), which honors the alignment
constraints, results in different offsets for the sections at higher
addresses than the extended section.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@3370 a27c51de-2d2b-4d1b-a6e8-f7f87814f453