ELF Rel is parsed
ELF Rel and ELF Rela sections are both represented by
SgAsmElfRelocSection containing a list of entries of type
SgAsmElfRelocEntry. Each entry corresponds to an Elf32_Rel or Elf32_Rela
struct (see "elf.h"), but stored in native byte order. The determination
of which format to use for all the entries of a relocation section depends
on the boolean "p_uses_addend" in the SgAsmElfRelocSection. The
SgAsmElfRelocEntry::p_r_addend data member will always be zero if its
section has p_uses_addend==false.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@3613 a27c51de-2d2b-4d1b-a6e8-f7f87814f453