ELF section symbols are not given names
They don't have names in the symbol table and generating a name during
parsing just causes that generated name to be written back out to the
file when unparsing. The name of these symbols can be obtained by
noticing that they refer to a section and then obtaining the section
name instead.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@3691 a27c51de-2d2b-4d1b-a6e8-f7f87814f453