Added the binaryAsyncSignalHandler test
This test parses a binary executable and looks for signal handlers. It
then tries to determine, by looking at the function call graph rooted at
each signal handler, whether a signal handler calls a dynamically linked
function that is not known to be async signal safe, either directly or
indirectly.

At this time it cannot handle stripped executables (because the
disassembler doesn't do a good enough job finding function boundaries).
Also, it only considers signal handlers that are set with a constant
second argument to signal (because we don't track constant
propogation), and we don't support sigaction yet.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@3909 a27c51de-2d2b-4d1b-a6e8-f7f87814f453