Added tests for finding function boundaries
If an executable is stripped (like most are) then there's no symbol
table describing ranges of addresses for each function. We need some
other way to detect functions so that we can still do function-based
analyses.

This new executable is meant to be a temporary place to experiment with
these methods, which will eventually augment the disassembler where it
creates the SgAsmFunctionDeclaration nodes.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@3910 a27c51de-2d2b-4d1b-a6e8-f7f87814f453