Rewrote isInstructionExternallyVisible()
The get_externallyVisible() function was recently removed from ROSE (by
me) because its intention was unclear. It wasn't used by anything in
ROSE and it seems to be superceded by the code that detects function
starts.

The new implementation checks to see whether the instruction is an entry
point of a function.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@4375 a27c51de-2d2b-4d1b-a6e8-f7f87814f453