Ability to detect signal handlers with sigaction
This works when sigaction is called with a struct siaction address which
is filled in with a signal handler address which is a constant or a
function argument.

It also works for a wrapper to which an already initialized sigaction
struct is passed, as long as the caller of the wrapper is the function
where the siaction struct is initialized.

See example code: test4.c test5.c test6.c

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@4501 a27c51de-2d2b-4d1b-a6e8-f7f87814f453