Reorgnized the instruction unparsing code
I decided it would be useful to be able to unparse the mnemonic
(actually, everything before the first operand) and the operands
seperately so that tools can have more control over how such things are
formatted into columns.  To that end, I created new functions:

  std::string unparseMnemonic(SgAsmInstruction*);
  std::string unparseExpression(SgAsmExpression*);

and rewrote unparseInstruction in terms of those two. I also made all
the architecture-specific unparsers private.

This change also fixes some bugs where the architecture-specific
functions made calls to one of the other architecture-specific functions
instead of its own functions.

In the future, time permitting, I may move these functions into the
SgAsmInstruction and SgAsmExpression classes and make the
architecture-specific helpers virtual so as to eliminate the "switch"
statements that are used to decide what class we're unparsing.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@5625 a27c51de-2d2b-4d1b-a6e8-f7f87814f453