Minor reformatting
I replaced some commented-out code with "#if 0" rather than C comments
because it was rather extensive, and doing it this way allows C tools
to still process the code as code.  It's questionable whether the
inactive code should even be there -- perhaps we should replace it with
a comment saying that it was removed in such-n-such a revision.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@5626 a27c51de-2d2b-4d1b-a6e8-f7f87814f453