Tool to expand core dumps
Linux core dumps don't contain data for their read-only segments. This
tool attempts to expand a core file by augmenting it with data from the
corresponding executable.

Also, a core dump ELF Header has a null entry address. We grab the entry
address from the executable.

Finally, we copy ELF Sections from the executable into the new core
file.  The previous behavior was to use the executable file sections to
initialize the contents of the core file segments. This change causes an ELF
Section Table and corresponding string table to be added to the core
file and the sections whose data was copied are also defined as ELF
Sections in the core file.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@5635 a27c51de-2d2b-4d1b-a6e8-f7f87814f453