Partial revert of r4507: fixed instruction semantics for rep_stosN
Reverts the x86InstructionSemantics.h from r4507 while preserving the
new code.  The old code semantics were to execute only one iteration of
the stos command, adjusting %eip so as to loop over the instruction.
This is how the CPU actually executes "rep" and a ptrace() single-step
demonstrates this behavior. On the other hand, the new code (introduced
in r4507) is needed by the asyncSignalSafety checker and executes the
entire loop in one shot.

The new and old code co-exist. The old code is in
X86InstructionSemantics while the new code was moved into
asyncSignalSafety.C. Various changes were needed in order be able to use
X86InstructionSemantics as a superclass.

This commit also fixes various bugs:
  * loop termination condition in pre-r4507 code
  * compile error in dead code of X86InstructionSemantics
  * inverted logic in equalToZero for verifySemantics.C

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@5949 a27c51de-2d2b-4d1b-a6e8-f7f87814f453