Improved testing for <rose.h>
If your build directory is located somewhere under the source directory
then the detect_roseh and detect_rosec rules in the top-level makefile
fail (except on a fresh build) because the process of building ROSE
creates some source files that violate this very rule.

A simple fix is to just not search directories that look like build
directories. This also speeds up the test substantially when you have
lots build directories under the source tree.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6032 a27c51de-2d2b-4d1b-a6e8-f7f87814f453