RTED now handles test c_I_2_a.c.
In this test, a pointer that refers to a block of memory is re-assigned, but not
via a malloc call.  In such cases, it is still necessary to check for memory
leaks (with calls to `checkMemoryLeakIssues`).  `RoseSystem_roseInitVariable`
now checks for memory issues whenever the assignee (i.e. the variable being
initialized) is of type SgPointerType and its current `address` is non-null.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6315 a27c51de-2d2b-4d1b-a6e8-f7f87814f453