RTED: Updated make check, et. al to work with exception mechanism, and did some work on the cstdlib wrapper.
The cpp runtimesystem now throws exceptions on violations, so the wrapper
functions have to try/catch them to maintain the previous semantics (i.e. that
we exit(1) when we don't find an error, and exit(0) otherwise).

Also the function call checking in the wrapper now delegates calls to the
appropriate function in the runtime system (e.g. check_strlen, check_memcpy).
Note that the string_errors tests do not yet all pass however.



git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6457 a27c51de-2d2b-4d1b-a6e8-f7f87814f453