RTED: worked on cstdlib checking.
Improved check_strcat and check_strncat so that they check for write errors from
the null terminator and not the start of the string.  This implies destination
needs to pass check_string, which is appropriate.

Also improved some of the output from mem. manager to give addresses in hex and
also give the size.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6465 a27c51de-2d2b-4d1b-a6e8-f7f87814f453