RTED: Memory manager now understands stack memory, and that freeing it is illegal.
A flag was added to MemoryType to indicate whether the memory is on the stack or
the heap.  Calls to freeMemory that do not explicitly state they are freeing
stack memory (i.e. calls from endScope) will fail if freeing memory flagged as
on the stack.  In calls to freeMemory that claim to be freeing stack memory, but
actually aren't, the flag is simply ignored, although this is probably wrong as
it's likely an error in the runtime system.

As a result of the above, we now pass more of the allocation_deallocation tests.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6546 a27c51de-2d2b-4d1b-a6e8-f7f87814f453