Improvements to emailed make output
If an error occurs deep within recursive makes, then we print only the
output from the inner-most make (at the leaf).  The old behaviour was to
print the output from the inner-most make and output from the makes
that are on the branch to the leaf where the error occurred (but
suppressing output from makes on other branches).

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6579 a27c51de-2d2b-4d1b-a6e8-f7f87814f453