Fixed x86 disassembler bug for REX.X with SIB index==4
The special handling for SIB.index==4 should only happen when REX.X is
zero since REX.X provides and extra bit for SIB.index, changing the 0x4
to an 0xC (0xC is register r12).

This bug manifested itself in the incorrect disassembly of instructions
like the following:

 4a 8d 44 23 ff: lea rax, [rbx + r12 - 1]

which was incorrectly disassembled having the second operand
"[rbx - 1]".

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6614 a27c51de-2d2b-4d1b-a6e8-f7f87814f453