Added an X86 Assembler
This work is not complete yet, but the assembler is able to correctly
reassemble all the executables in my 64-bit /bin directory.

Top-level assembly class is "Assembler" with subclasses for each
architecture dependency (same layout as for the disassembler).

The x86 assembler is rule-driven and the rule table is populated by a
script in the src/frontend/Disassemblers directory which reads the Intel
Instruction Set Reference PDF documentation and generates
AssemblerX86Init.[Ch].

The X86InstructionEnum (list of each "kind" of x86 instruction) is now
machine generated and cross referenced to the Intel Instruction Set
Reference. There's also a to_str() function that returns the string
representation of the enumerated constants.

Also enhanced the SgAsmExecutableFileFormat::hexdump methods by making
their output more flexible. The cut-and-pasted version for printing
instructions can now be removed.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6620 a27c51de-2d2b-4d1b-a6e8-f7f87814f453