x86 disassembler returns correct data types
The data types for x86 instruction operands are consistent with the
encoding of those operands in the machine instructions. For instance, if
an instruction has an "imm8" (8-bit immediate) operand then the
disassembler will create a SgAsmByteValueExpression rather than a
SgAsmQuadWordValueExpression.

Also updated test results since the string representation of an assembly
instruction changed slightly due to the smaller values returned by the
disassembler.  For instance:
   add eax, 0xffffff80
becomes
   add eax, 0x80
if the 0x80 was stored as a single byte in the machine code.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6646 a27c51de-2d2b-4d1b-a6e8-f7f87814f453