Fine tuned SEARCH_FOLLOWING mode for disassembler
This mode of disassembly causes the disassembler to start disassembling
after each basic block, regardless of whether the address is a known
successor (for instance, disassemble even after a HLT instruction).

The fine tuning is that the address is not disassembled if it falls
outside the mapping from virtual address to file offset (such as when
the address is past the end of the file or within a non-executable
section).  This change makes SEARCH_FOLLOWING behave more like the other
search methods.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6651 a27c51de-2d2b-4d1b-a6e8-f7f87814f453