Added better x86 prefix matching
When the assembler is operating in a mode where it's trying to generate
an encoding that matches the original bytes used to disassemble an
instruction (i.e., Assembler::get_encoding_type()==ET_MATCHES), it tries
harder to generate x86 instruction prefixes in the same order and
repetition as the original encoding.

This fixes problems encountered often with NOP instructions where the
compiler has padded the NOP encoding to make it a particular size. Some
compilers add 0x66 (Operand Size Override) prefixes, others insert
segment overrides, others change operand sizes. But the fix is not
particular to NOPs--it applies equally to all kinds of x86 instructions.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6652 a27c51de-2d2b-4d1b-a6e8-f7f87814f453