RTED: Tweak extraction scripts to treat all memory leaks as end-of-program errors.
We now pass the memory_leaks tests, but not very well.  In particular, we don't
identify the line number that causes the leak, we just identify that a leak
occured at the end of the program.

I believe that passing these tests properly will require dataflow analysis in
the RTS to identify those addresses that are "leaked" (i.e. not pointed to by
anywhere in memory, nor reliably computable from somewhere in memory).



git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6665 a27c51de-2d2b-4d1b-a6e8-f7f87814f453