NMI: Added nmi-postmortem script.
nmi-postmortem can be run on a submission host (at present, only s005) with a
run-id to copy and extract results.tar.gz to the run machine and setup a
source-able script .nmi-postmortem, which will source the run environment and cd
the user to the userdir.  The script ends by exec-sshing to the run machine and
sourcing .nmi-postmortem (indirectly, via bashrc).

Example Usage:
	./nmi-postmortem 169831


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6694 a27c51de-2d2b-4d1b-a6e8-f7f87814f453