RTED: Improvements towards passing array bounds checks.  Array types in classes now registered properly, plus some misc. cleanup.
We still don't pass all of the array bounds checks.  Presently we fail on
multi-dimensional arrays, but much of the infrastructure for dealing with them
is in place.  The instrumentation and wrapper needs to determine and pass the
indirection level for types with base types (i.e. pointers and arrays).

*	When PointerManager creates array type pseudo-pointers, they're immediately
	set to the source address.
*	RsArrayType's second argument, size, now actually treats the arg as size
	rather than element count.
*	Class registration instrumentation now adds the size of members (needed for
	SgArrayType), as well as their base type.

*	tests was removed as a subdir, along with its Makefile.
*	removed some dead code.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6696 a27c51de-2d2b-4d1b-a6e8-f7f87814f453