RTED: Some work with Thomas Heller on supporting the C++ tests in a maintainable way.
The basic approach is in place, which is to have cross-dependencies on all the
rose c[pp] files with the corresponding rted c[pp] files (so the .c files depend
on their non-existent .cpp rted files) with a rule that happily ignores the fake
dependencies.

Some of the convenience rules need to be copied/expanded for the C++ tests,
however.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6711 a27c51de-2d2b-4d1b-a6e8-f7f87814f453