RTED: Progress on memory leaks tests.  Now all but two pass, and those two fail because of int** handling, just as the array_index_out_of_bounds tests that fail (e.g. c_E_5_1_a).
Now we produce compilable code for calls to variable access that result from
expressions with pointer arithmetic (e.g. sum += *(p++), from which we would
previously produce expressions like &(p++)).

The expected error extraction is more precise about the cases for which it
simply expects a memory leak to be discovered when the program terminates.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6744 a27c51de-2d2b-4d1b-a6e8-f7f87814f453