RTED: Work on memory leak errors, support for double (or n-level) pointers.
PointerManager will now check memory leaks on pointer deletion, which catches
some leaks from implied pointers (.e.g with `int **x`, *x is a pointer, but not
a variable).  This will give false positives for "computable pointers", which we
need data flow analysis to avoid.

Some work on tweaking scope code so that the error line numbers match better,
and a couple of helper tasks added to the makefile, namely:
	run_log/c
	run_log/cpp


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6798 a27c51de-2d2b-4d1b-a6e8-f7f87814f453