NMI: nmi-postmortem now automatically sets up home dir on run host. Documentation improved.
When nmi-postmortem is invoked, it will check a file, .nmi-hosts, to see if the
run host is known to have been set up.  If this is not the case, nmi-postmortem
will now setup the home directory on the run host (no effect if this has already
been done) and copy the (now-passphraseless) public key to the run hosts's
authorized_keys file.

Because the public key now has an empty passphrase, having ssh-agent running
(and its associated check in nmi-postmortem) is no longer necessary.

The documentation has been updated to make some things more explicit.


git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6911 a27c51de-2d2b-4d1b-a6e8-f7f87814f453