ExecFormats: referenced extents tracked at file level
The old behavior was that each section tracked what part of the file was
read. The benefit of the old approach was that if the section was
relocated the referenced/unreferenced list still applied.

The new method is that references are tracked at the file level, which
will work better when we start using RvaFileMap to read from the file
based on virtual addresses.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6976 a27c51de-2d2b-4d1b-a6e8-f7f87814f453