ExecFormats: readers throw exceptions
The default action for all the read_content*() functions is to throw an
exception (either SgAsmExecutableFileFormat::ShortRead for direct file
reads or RvaFileMap::NotMapped for virtual memory) if an attempt is made
to read data which doesn't exist.

These functions all take an optional final argument that toggles between
throwing the exception and zero-padding the result.  In the latter case,
the return value indicates how many bytes were actually read.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6984 a27c51de-2d2b-4d1b-a6e8-f7f87814f453