ExecFormats: rose_*.s files were empty
Fixes a bug where the assembly dump was not being written to the
name specified in the SgBinaryFile node, but only to the *.dump
file produced by the SgAsmGenericFile node.

Note that the assembly in the rose_*.s file contains output from all
SgAsmInterpretation nodes.

This hopefully fixes the recent problem with the tutorial.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@7473 a27c51de-2d2b-4d1b-a6e8-f7f87814f453