ExecFormats: interpretations point to multiple files
An SgAsmInterpretation now points to a list of SgAsmGenericHeader
objects in order to handle dynamic linking or other situations where a
single memory map (e.g., for disassembly) would point into multiple
binary files.

The Loader classes (for creating memory maps) were changed slightly so
that a map spanning multiple files can be built incrementally.  The map
is now stored in the SgAsmInterpretation node rather than in the
SgAsmGenericHeader.  (The PE parser uses it's own map for parsing.)

LibraryIdentification::generateOpCodeVector() makes even less sense now
than it did before. Not only is the memory mapping discontinuous in the
file, but it can now span multiple files.

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@7622 a27c51de-2d2b-4d1b-a6e8-f7f87814f453