Added a new test for testing the correctness of the sizeof operation on different compilers against the EDG front-end.
This test case tested the sizeof operation on different C/C++ data types, including build-in and user defined types. The tests have been conducted on GNU and Intel compile. The results show that GNU compiler behaved differently with EDG front-end on processing nested #pragma pack directives. But Intel compiler is always consistent with EDG front-end. The reason is included in the README file.



git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@7862 a27c51de-2d2b-4d1b-a6e8-f7f87814f453