Assembler: ARM operand unparser can be invoked from arbitrary node
The ARM instruction operand unparser (that generates ARM assembly from
a SgAsmExpression) can be invoked from arbitrary nodes in the expression
tree rather than just the top node.  Note that the string returned by
invoking it in the middle of an expression will probably not be valid
ARM assembly.

For instance, invoking the unparser at the SgAsmBinarySubtractPreupdate
node of the following instruction:

    str r14, [r13, #-4]!

will result in the string:

    "r13, #-4" and "!"

git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@7988 a27c51de-2d2b-4d1b-a6e8-f7f87814f453