New tool to show effectiveness of function detection
The testFunctionDetection.pl perl script runs the detectFunctions tool
on the original binary and a stripped version and then shows what
functions, if any, were detected in the former but not the latter.

Also adds an instruction pattern to the mark_func_patterns() in the
Partitioner class, which previously was a no-op.