Moved instruction semantics verifier to test directories
This is to cause it to be distributed so that external developers that
work on instruction semantics have a way to test their work.  The test
is not run automatically because it requires a fair degree of human
intervention to set up a slave machine for the architecture being
tested.