X86 assembly constant value format
Constants in the assembly output for x86 were ambiguous. There was no
way to tell whether an operand like

   DWORD PTR ss:[ebx + 0xfc]

referred to a memory location whose address was ebx+252 or ebx-4 because
it was impossible to tell whether 0xfc was an 8- or 32-bit constant.

The new output format always prints constants using a fixed number of
hexadecimal digits, so 0xfc is an eight-bit value while 0x000000fc is a
32-bit value.