Implemented smgit2-pull
smgit2-pull is not as simple as just running git-pull in each
submodule because (1) we need to avoid conflicts in the main
repo due to differing submodule SHA1s that result from the
merge in the submodules, (2) we need to be aware that the
branch names in the submodules are not necessarily the same
as the branch names in the main repo.

I don't normally use git-pull, so I'm not positive this is
set up correctly. Therefore, use of smgit2-pull will require
user confirmation before the merge command is run.

A better mode of operation is to use git-fetch, then inspect
the changes, then git-merge if desired.  The git-pull command
is more like Subversion's "update" in that it combines the
fetch and merge without giving the user a chance to see the
changes beforehand (at least Git allows you to reset after
a bad merge).