Removed smgit2 dependency on Getopt::Long
This script depends on a version of Getopt::Long that can parse
command-line switches from an array. This version of Getopt::Long
is not installed on all the machines where we want to use it, yet
we don't want users to have to install it.

Therefore, the correct version of Getopt::Long has been inlined
into smgit2.