Minor changes to the way disassemblers are registered.
The Disassembler and subclasses thereof no longer use factories
for their creation.  Instead, specific disassemblers are registered
with the Disassembler class and a lookup() class method returns
a disassembler that can handle the specified machine code.

A clone() virtual method was added so that if a user wants to
modify some settings for a disassembler, such as how it searches
for instructions to disassemble, he can do so without modifying
the disassembler that was originally registered.