Added SgAsmFunctionDeclaration::FUNC_LEFTOVERS
This bit is added to the function reason bit vector when the function
is created by the partitioner to hold basic blocks that could not be
associated with any other function. These basic blocks are often a
result of an overly aggressive disassembly.

This bit is also one of the search heuristic switches for the
Partitioner class and is enabled by default. When disabled, the
partitioner will discard all basic blocks that could not be
assigned to a function (these blocks will not appear in the AST).