Tweaking instruction partitioner
While assigning instructions to blocks and blocks to functions,
if the partitioner splits a block (due to discovering a new
control flow edge) then it temporarily abandons discovery for
the affected function and starts again.  It does this because
splitting a basic block might cause the control flow edges
to change even though we've processed the old edges.