smgit2 handles case where merge deletes all submodules
For instance, if the EDG submodule is removed from a branch and
then you try to merge from that branch, smgit2 will issue a
warning about the submodule disappearing, but will operate
correctly.

It actually was correct before also, it just spit out a bunch
of junk messages from running git commands that didn't make
sense in that situation.