Added function hashing to disassembler test.
Computes function hash as the XOR of the unique hashes of
the function's basic blocks.  The hash of the basic block
is computed as before, namely, the SHA1 digest of the
semantic description of the basic block.

To see output, run "disassemble --show-functions --quiet a.out"
in the tests/roseTests/binaryTests directory, where a.out
is any 32-bit x86 executable.