Added a debugging name to MemoryMap::MapElement
The name is intended to be used to aid in debugging since there
are situation where the information will not be accurate (for instance,
when map elements are removed and added but the removals do not
use the same address/size pairs as the original insertion.

The MemoryMap::dump() method prints this name.

Memory map debug names are limited to 35 characters