Partial revert of ad3e21041fceaf9799728924dd623029e9cc6bfb w/comment
Dan had commented out CheckIsModifiedFlagSupport() which broke binary
unparsing.  The unparser requires knowledge about whether the AST
has been modified after the call to frontend() and before the call
to backend().  If the tree is modified then ROSE gives sections a
chance to reallocate themselves in the file/memory map before unparsing.
This can sometimes have undesirable effects if the tree was not
modified, so we avoid the reallocation in that case.

Dan had also (inadvertently?) commented out all but a couple of
the tests in tests/roseTests/binaryTests, which this revert
restores.