The "disassemble" tool accepts mapping permissions on the cmd line
When disassembling a raw buffer such as a memory dump, one invokes
"disassembler" with the "--raw=ENTRY_VA" switch and provides pairs
of non-switch arguments, each containing a file name and mapping
address.

I've extended the mapping address so that it may be suffixed with the
letters 'r', 'w', and/or 'x' for read, write, and execute.