MemoryMap read/write methods check map permissions.
Extra argument to read/write methods indicate what mapping permissions
are required for the operation.  The defaults are MM_PROT_READ for read()
and MM_PROT_WRITE for write().

The old behavior can be obtained by specifying a final argument
of MM_PROT_NONE, which causes the read/write to succeed provided
the memory is actually mapped.